requirements.txt
streamlit
streamlit_chat
pandasai
python-dotenv
tabulate
langchain

